# Toward better multi-threading performance