--- a/Core/Instrument/DetectorMask.h
+++ b/Core/Instrument/DetectorMask.h
@@ -53,7 +53,7 @@ public:
     void removeMasks();
 
     //! returns true if has masks
-    bool hasMasks() const { return numberOfMasks()>0; }
+    bool hasMasks() const { return !m_shapes.empty(); }
 
     int numberOfMaskedChannels() const { return m_number_of_masked_channels; }