--- a/Core/Instrument/SphericalDetector.cpp
+++ b/Core/Instrument/SphericalDetector.cpp
@@ -86,18 +86,17 @@ size_t SphericalDetector::getIndexOfSpecular(const Beam& beam) const
     return totalSize();
 }
 
-SphericalPixel::SphericalPixel(Bin1D alpha_bin, Bin1D phi_bin)
+SphericalPixel::SphericalPixel(const Bin1D& alpha_bin, const Bin1D& phi_bin)
     : m_alpha(alpha_bin.m_lower), m_phi(phi_bin.m_lower),
       m_dalpha(alpha_bin.getBinSize()), m_dphi(phi_bin.getBinSize())
 {
-    m_solid_angle = std::abs(m_dphi*(std::sin(m_alpha+m_dalpha) - std::sin(m_alpha)));
+    auto solid_angle_value = std::abs(m_dphi*(std::sin(m_alpha+m_dalpha) - std::sin(m_alpha)));
+    m_solid_angle = solid_angle_value <= 0.0 ? 1.0 : solid_angle_value;
 }
 
 SphericalPixel* SphericalPixel::clone() const
 {
-    Bin1D alpha_bin(m_alpha, m_alpha+m_dalpha);
-    Bin1D phi_bin(m_phi, m_phi+m_dphi);
-    return new SphericalPixel(alpha_bin, phi_bin);
+    return new SphericalPixel(*this);
 }
 
 SphericalPixel* SphericalPixel::createZeroSizePixel(double x, double y) const
@@ -125,6 +124,5 @@ double SphericalPixel::getIntegrationFactor(double /* x */, double y) const
 
 double SphericalPixel::getSolidAngle() const
 {
-    if (m_solid_angle<=0.0) return 1.0;
     return m_solid_angle;
 }