--- a/Core/Instrument/SphericalDetector.h
+++ b/Core/Instrument/SphericalDetector.h
@@ -64,8 +64,7 @@ protected:
 class SphericalPixel : public IPixel
 {
 public:
-    SphericalPixel(Bin1D alpha_bin, Bin1D phi_bin);
-    virtual ~SphericalPixel() {}
+    SphericalPixel(const Bin1D& alpha_bin, const Bin1D& phi_bin);
 
     SphericalPixel* clone() const override;
     SphericalPixel* createZeroSizePixel(double x, double y) const override;