… Core/Instrument/DetectorElement.cpp
… Core/Instrument/DetectorElement.h
… Core/Simulation/OffSpecSimulation.cpp
… Core/Simulation/Simulation2D.h
… Core/SimulationElement/SimulationElement.cpp
… Core/SimulationElement/SimulationElement.h
… Tests/Functional/Core/CoreSpecial/CoreSpecialTestFactory.cpp
… Tests/Functional/Core/CoreSpecial/DetectorTest.cpp
… Tests/Functional/Core/CoreSpecial/DetectorTest.h
… Tests/Functional/Core/CoreSpecial/MultiThreadPerformanceTest.cpp
… Tests/Functional/Core/CoreSpecial/MultiThreadPerformanceTest.h
… Tests/Functional/Core/CoreSpecial/MultiThreadPerformanceTestComponents.cpp
… Tests/Functional/Core/CoreSpecial/MultiThreadPerformanceTestComponents.h
… Tests/UnitTests/Core/Detector/RectangularDetectorTest.cpp
… Tests/UnitTests/Core/SimulationElement/SimulationElementTest.cpp
… auto/Wrap/doxygen_core.i
… auto/Wrap/doxygen_fit.i
… auto/Wrap/libBornAgainCore.py
… auto/Wrap/libBornAgainFit.py
… auto/Wrap/libBornAgainFit_wrap.cpp