This pull request contains a number of improvements toward better performance in a multi-thread environment: 

+ Functional test to measure the performance in different scenarios.
+ Faster SimulationElement construction.
+ Some steps toward disantangling DetectorElements and SimulationElements.

It provides 10%-30% better performance in a single-thread mode in various scenarios. Poor scaling with the number of threads mentioned in #914 is not improved yet.

Functional test introduces the following simulation configurations:
```
Simple     : Simple sample, small detector (300 times)
Real       : Simple sample, realistic detector (10 times)
Heavy      : Complex sample, realistic detector (1 time)
Large      : Trivial sample, huge detector (1 time)
Wave       : Trivial sample, tiny detector, huge wavelength distribution (1 time)
MC         : Complex sample, small detector, monte-carlo integration (1 time)
```
Before modifications:
```
Performance in msec and thread scaling efficiency for various simulations.
n.t.  | Simple        | Real          | Heavy         | Large         | Wave          | MC            |
1     | 4882   100.0  | 6027   100.0  | 9691   100.0  | 4184   100.0  | 3726   100.0  | 8667   100.0  | 
2     | 7020   34.8   | 4406   68.4   | 4926   98.4   | 2852   73.4   | 6125   30.4   | 4419   98.1   | 
4     | 4905   24.9   | 3104   48.5   | 2657   91.2   | 2200   47.5   | 4928   18.9   | 2372   91.3   | 
6     | 4353   18.7   | 2683   37.4   | 1881   85.9   | 1978   35.3   | 5103   12.2   | 1612   89.6   | 
8     | 3800   16.1   | 2341   32.2   | 1443   83.9   | 1840   28.4   | 4291   10.9   | 1222   88.7   | 
10    | 3806   12.8   | 2164   27.9   | 1183   81.9   | 1785   23.4   | 4315   8.6    | 980    88.4   | 
12    | 3539   11.5   | 2099   23.9   | 1048   77.1   | 1727   20.2   | 4377   7.1    | 850    85.0   | 
14    | 3129   11.1   | 1989   21.6   | 942    73.5   | 1718   17.4   | 4332   6.1    | 736    84.1   | 
16    | 3131   9.7    | 1908   19.7   | 853    71.0   | 1697   15.4   | 4434   5.3    | 671    80.7   | 
20    | 3531   6.9    | 1922   15.7   | 905    53.5   | 1728   12.1   | 5296   3.5    | 736    58.9   | 
24    | 3570   5.7    | 1904   13.2   | 841    48.0   | 1719   10.1   | 5799   2.7    | 639    56.5   | 
28    | 3626   4.8    | 1741   12.4   | 816    42.4   | 1705   8.8    | 6054   2.2    | 605    51.2   | 
32    | 3715   4.1    | 1683   11.2   | 764    39.6   | 1696   7.7    | 6403   1.8    | 576    47.0   | 
```

After modifications:

```
Performance in msec and thread scaling efficiency for various simulations.
n.t.  | Simple        | Real          | Heavy         | Large         | Wave          | MC            |
1     | 4235   100.0  | 6018   100.0  | 9061   100.0  | 3394   100.0  | 2885   100.0  | 8253   100.0  | 
2     | 6411   33.0   | 4269   70.5   | 4654   97.3   | 2219   76.5   | 4988   28.9   | 4329   95.3   | 
4     | 4440   23.8   | 3080   48.8   | 2501   90.6   | 1711   49.6   | 4106   17.6   | 2231   92.5   | 
6     | 3916   18.0   | 2628   38.2   | 1724   87.6   | 1518   37.3   | 4144   11.6   | 1491   92.3   | 
8     | 3589   14.7   | 2230   33.7   | 1344   84.3   | 1407   30.2   | 3636   9.9    | 1150   89.7   | 
10    | 3343   12.7   | 2031   29.6   | 1111   81.6   | 1378   24.6   | 3623   8.0    | 940    87.8   | 
12    | 3131   11.3   | 1972   25.4   | 989    76.3   | 1341   21.1   | 3722   6.5    | 802    85.8   | 
14    | 2989   10.1   | 1841   23.3   | 870    74.4   | 1326   18.3   | 3697   5.6    | 725    81.3   | 
16    | 3027   8.7    | 1812   20.8   | 800    70.8   | 1293   16.4   | 3780   4.8    | 638    80.8   | 
20    | 3397   6.2    | 1813   16.6   | 856    52.9   | 1331   12.7   | 4631   3.1    | 696    59.3   | 
24    | 3439   5.1    | 1753   14.3   | 774    48.8   | 1302   10.9   | 5050   2.4    | 605    56.8   | 
28    | 3490   4.3    | 1641   13.1   | 732    44.2   | 1305   9.3    | 5255   2.0    | 596    49.5   | 
32    | 3606   3.7    | 1565   12.0   | 692    40.9   | 1312   8.1    | 5538   1.6    | 545    47.3   | 
```

More developer's dettails [here](http://apps.jcns.fz-juelich.de/redmine/issues/2433)