@juanmcloaiza 2020-04-14
> Some steps toward disantangling DetectorElements and SimulationElements.

Nice, this entanglement of Simulation-Detector-Instrument is something that I wanted to comment on, but didn't dare.

Here's a plot about the timings reported on this PR, I hope it is useful:

![image](https://user-images.githubusercontent.com/14291849/79221456-24ecb000-7e56-11ea-9659-ca9e1fbf6697.png)